pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "find-domains"
version = "0.1.0"
description = "Discover valuable typosquatting domain opportunities using trend intelligence and AI analysis"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pytrends>=4.9",
    "openai>=1.0",
    "httpx>=0.27",
    "dnspython>=2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
find-domains = "find_domains.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/find_domains"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
